Add basic exception feature
- class Exception and its derived class
- use basic REPL in main
- fix wrong `size_t` alias (uint -> size_t)
- fix `show` function default argument bug

+++ defs.hpp
@@ -9,7 +9,7 @@
 
 namespace emehcs {
 
-using uint = ::std::size_t;
+using size_t = ::std::size_t;
 
 constexpr ::std::string_view Symbol = "!$%&|*+-/:<=?>@^_~#";
